Database addition/new preferences
Now uses Firebase realtime database to store urls of all uploaded images. This is used to display images in web portal, and for deleting all images when deleting account.
Also reworked userEmail/loggedIn to be saved as preferences when user is logged in. This reduces number of args passed to each activity, instead fetches this data from preferences.
Fixed @Override javadoc issue.